backend/requirements.txt
asn1crypto==1.2.0
astroid==2.3.2
cached-property==1.5.1
cffi==1.13.1
colorama==0.4.1
cryptography==2.8
Django==2.2.6
django-cors-headers==3.1.1
django-environ==0.4.5
django-extensions==2.2.5
django-filter==2.2.0
django-organizations==1.1.2
django-smart-selects==1.5.4
django-url-filter==0.3.13
djangorestframework==3.10.3
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
enum-compat==0.0.3
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
mypy==0.740
mypy-extensions==0.4.3
numpy==1.17.3
Pillow==6.2.1
psycopg2-binary==2.8.4
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.3
python-magic==0.4.15
pytz==2019.3
PyYAML==5.1.2
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
wrapt==1.11.2
